+++ apps/api/filters/annotations.py
@@ -4,9 +4,9 @@ from apps.core.models import Annotation, AnnotationItem
 
 
 class AnnotationFilter(django_filters.FilterSet):
-    user_acquisition_id = django_filters.UUIDFilter(lookup_expr="iexact")
+    user_acquisition_id = django_filters.UUIDFilter(lookup_expr="exact")
     title = django_filters.CharFilter(lookup_expr="icontains")
-    user_id = django_filters.UUIDFilter(lookup_expr="iexact", field_name="user_acquisition__user_id")
+    user_id = django_filters.UUIDFilter(lookup_expr="exact", field_name="user_acquisition__user_id")
 
     class Meta:
         model = Annotation
@@ -26,7 +26,7 @@ class AnnotationFilter(django_filters.FilterSet):
 
 
 class AnnotationItemFilter(django_filters.FilterSet):
-    annotation_id = django_filters.UUIDFilter(lookup_expr="iexact")
+    annotation_id = django_filters.UUIDFilter(lookup_expr="exact")
     page_number = django_filters.NumberFilter(field_name="page")
 
     class Meta: