+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog suppose to follow rules defined in the [changelog.md](https://cha
 - **Changed**: Sort entries in Complete OPDS catalog by `created_at`
 - **Changed**: `EntrySerializer` now returns URL path in the `thumbnail` property instead of base64 image (performance)
 - **Fixed**: `page` -> `page_number` in AnnotationItemFilter
+- **Fixed**: `exact` lookup fields on ForeignKeys
 
 ## 0.8.0 : 2024-05-09